rename either -> common
